refactor(models): remove services/config dependencies
Move Pydantic model definitions and BranchConvention to models/ as canonical
locations, extract impure methods into standalone helpers in config/.

Changes:
- models/orchestra_config.py: Replace re-export with actual Pydantic models
  (removed get_handoff_state_label, get_manager_usernames, get_supervisor_prompt)
- config/orchestra_config.py: Convert to re-export + 3 standalone helper functions
  (get_handoff_state_label, get_manager_usernames, get_supervisor_prompt)
- models/branch_convention.py: Move BranchConvention dataclass here (was in config/)
- config/branch_convention.py: Convert to re-export shim
- Update 11 call sites: instance methods → standalone function calls
- Update 5 test files: mock standalone functions instead of instance methods

Architecture verification:
- No models files import from services/ or config/
- All tests pass
- mypy: 0 errors
- ruff: all checks passed

Refs: #1247
